Move SubmoduleCollection to submodule.py
